refactorings
- Rename logging classes
- Rename Logger to TypedLogger
- Simplify logging config
- Reformat imports
- Introduce ShareTrace reference config

+++ src/main/java/io/sharetrace/util/logging/metric/BaseGraphEccentricity.java
@@ -5,9 +5,9 @@ import org.immutables.value.Value;
 @Value.Immutable
 interface BaseGraphEccentricity extends LoggableMetric {
 
-  int radius();
+  long radius();
 
-  int diameter();
+  long diameter();
 
   long center();
